# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  model.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      pagePath:
        type: string
      path:
        type: string
      pid:
        type: string
    type: object
  model.News:
    properties:
      author:
        type: string
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.Person:
    properties:
      address:
        type: string
      age:
        type: string
      birthday:
        type: string
      hometown:
        type: string
      id:
        type: integer
      identity:
        type: string
      name:
        type: string
      sex:
        type: string
      userId:
        type: integer
    required:
    - address
    - age
    - birthday
    - hometown
    - identity
    - name
    - sex
    - userId
    type: object
...
host: localhost:9090
info:
  contact:
    email: grezkl@protonmail.com
    name: grezkl
    url: http://github.com/grezkl
  description: e-gov server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: e-gov server
  version: "1.0"
paths:
...
  /delFileBatch:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 文件编号数组
        in: path
        items:
          type: integer
        name: ids
        required: true
        type: array
      responses: {}
      summary: 删除文件（批量）
      tags:
      - File
  /delMenuBatch/:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 菜单编号数组
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 删除菜单（批量）
      tags:
      - Menu
  /delNewsBatch/:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 新闻编号数组
        in: path
        items:
          type: integer
        name: id
        required: true
        type: array
      responses: {}
      summary: 删除新闻（批量）
      tags:
      - News
...
  /deleteFileById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 文件编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 删除文件（单个）
      tags:
      - File
  /deleteMenuById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 菜单编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 删除菜单（单个）
      tags:
      - Menu
  /deleteNewsById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 新闻编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 删除新闻（单个）
      tags:
      - News
...
  /downloadFile/:name:
    get:
      consumes:
      - application/json
      parameters:
      - description: 文件名称
        in: path
        name: name
        required: true
        type: integer
      responses: {}
      summary: 文件下载
      tags:
      - File
...
  /filePage:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: createdAt
        type: string
      - in: query
        name: enable
        type: boolean
      - in: query
        name: id
        type: integer
      - in: query
        name: md5
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: size
        type: integer
      - in: query
        name: type
        type: string
      - in: query
        name: updatedAt
        type: string
      - in: query
        name: url
        type: string
      - in: query
        name: userId
        type: integer
      responses: {}
      summary: 更新文件信息
      tags:
      - File
...
  /menuPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 政务申请名称
        in: query
        name: name
        type: string
      responses: {}
      summary: 菜单管理分页
      tags:
      - Menu
  /newsPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 新闻名称
        in: query
        name: name
        type: string
      responses: {}
      summary: 新闻动态分页
      tags:
      - News
...
  /pullNewsById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 新闻编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 根据编号获取新闻
      tags:
      - News
...
  /pullPersonById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 根据用户编号获取实名信息
      tags:
      - Person
  /pullPersonByIdentity/{idt}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 身份证号
        in: path
        name: idt
        required: true
        type: string
      responses: {}
      summary: 根据身份证号获取真实信息
      tags:
      - Person
...
  /saveMenu:
    post:
      consumes:
      - application/json
      parameters:
      - description: 菜单内容
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/model.Menu'
      responses: {}
      summary: 保存菜单内容
      tags:
      - Menu
  /savePerson:
    post:
      consumes:
      - application/json
      parameters:
      - description: 保存实名信息
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/model.Person'
      responses: {}
      summary: 保存实名信息
      tags:
      - Person
...
  /updateNews:
    post:
      consumes:
      - application/json
      parameters:
      - description: 新闻内容
        in: body
        name: news
        required: true
        schema:
          $ref: '#/definitions/model.News'
      responses: {}
      summary: 保存新闻内容
      tags:
      - News
...
  /uploadFile:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      responses: {}
      summary: 文件上传
      tags:
      - File
...
swagger: "2.0"